+++ components/utils/src/lib.rs
@@ -4,6 +4,7 @@ extern crate errors;
 #[cfg(test)]
 extern crate tempdir;
 extern crate tera;
+extern crate walkdir;
 
 pub mod fs;
 pub mod site;