+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 to the public directory
 - Do not require themes to have a static folder
 - Now supports indented Sass syntax
+- Add search index building
 
 ## 0.3.2 (2018-03-05)